发现界面列表API数据接入和投票API数据接入

+++ living/living/Model/Request/Find/LMFindListRequest.h
@@ -10,6 +10,6 @@
 
 @interface LMFindListRequest : FitBaseRequest
 
--(id)initWithPageIndex:(int)pageIndex andPageSize:(int)pageSize;
+-(id)initWithPageIndex:(NSInteger)pageIndex andPageSize:(NSInteger)pageSize;
 
 @end
